/!\ Transform RegUsageMapper into a trait;
This allows several implementations of RegUsageMapper to co-exist
without the client being aware of it.

This also renames the existing RegUsageMapper to VrangeRegUsageMapper,
since it tracks the mappings for the whole lifetime of given virtual
ranges.

A few functions need to get a type parameter, and they get monomorphized
at compile-time.
